requirements.txt
fastapi 
uvicorn 
python-dotenv 
openai 
pandas 
openpyxl 
